gradeservice/requirements-base.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements-base.in
#
alabaster==0.7.13
    # via sphinx
anyio==3.7.0
    # via
    #   starlette
    #   watchfiles
appnope==0.1.3
    # via ipython
asttokens==2.2.1
    # via stack-data
attrs==23.1.0
    # via jsonschema
babel==2.12.1
    # via sphinx
backcall==0.2.0
    # via ipython
cachetools==5.3.1
    # via google-auth
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   otter-grader
    #   typer
    #   uvicorn
decorator==5.1.1
    # via ipython
dill==0.3.6
    # via otter-grader
docutils==0.20.1
    # via
    #   fica
    #   sphinx
executing==1.2.0
    # via stack-data
fastapi==0.97.0
    # via -r requirements-base.in
fastjsonschema==2.17.1
    # via nbformat
fica==0.2.2
    # via otter-grader
google-api-core==2.11.1
    # via google-api-python-client
google-api-python-client==2.89.0
    # via otter-grader
google-auth==2.20.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth-oauthlib==1.0.0
    # via otter-grader
googleapis-common-protos==1.59.1
    # via google-api-core
h11==0.14.0
    # via uvicorn
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.5.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
imagesize==1.4.1
    # via sphinx
ipython==8.14.0
    # via otter-grader
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   otter-grader
    #   sphinx
jsonschema==4.17.3
    # via nbformat
jupyter-core==5.3.1
    # via nbformat
jupytext==1.14.6
    # via otter-grader
markdown-it-py==2.2.0
    # via
    #   jupytext
    #   mdit-py-plugins
markupsafe==2.1.3
    # via jinja2
matplotlib-inline==0.1.6
    # via ipython
mdit-py-plugins==0.4.0
    # via jupytext
mdurl==0.1.2
    # via markdown-it-py
nbformat==5.9.0
    # via
    #   jupytext
    #   otter-grader
numpy==1.25.0
    # via pandas
oauthlib==3.2.2
    # via requests-oauthlib
otter-grader==4.3.3
    # via -r requirements-base.in
packaging==23.1
    # via sphinx
pandas==2.0.2
    # via otter-grader
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==3.6.0
    # via jupyter-core
prompt-toolkit==3.0.38
    # via ipython
protobuf==4.23.3
    # via
    #   google-api-core
    #   googleapis-common-protos
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydantic==1.10.9
    # via
    #   fastapi
    #   python-on-whales
pygments==2.15.1
    # via
    #   ipython
    #   sphinx
pyjwt==2.7.0
    # via -r requirements-base.in
pyparsing==3.0.9
    # via httplib2
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via uvicorn
python-multipart==0.0.6
    # via -r requirements-base.in
python-on-whales==0.61.0
    # via otter-grader
pytz==2023.3
    # via pandas
pyyaml==6.0
    # via
    #   fica
    #   jupytext
    #   otter-grader
    #   uvicorn
requests==2.31.0
    # via
    #   google-api-core
    #   otter-grader
    #   python-on-whales
    #   requests-oauthlib
    #   sphinx
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   asttokens
    #   google-auth
    #   google-auth-httplib2
    #   otter-grader
    #   python-dateutil
sniffio==1.3.0
    # via anyio
snowballstemmer==2.2.0
    # via sphinx
sphinx==7.0.1
    # via fica
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
stack-data==0.6.2
    # via ipython
starlette==0.27.0
    # via fastapi
toml==0.10.2
    # via jupytext
tqdm==4.65.0
    # via python-on-whales
traitlets==5.9.0
    # via
    #   ipython
    #   jupyter-core
    #   matplotlib-inline
    #   nbformat
typer==0.9.0
    # via python-on-whales
typing-extensions==4.6.3
    # via
    #   pydantic
    #   python-on-whales
    #   typer
tzdata==2023.3
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
uvicorn[standard]==0.22.0
    # via -r requirements-base.in
uvloop==0.17.0
    # via uvicorn
watchfiles==0.19.0
    # via uvicorn
wcwidth==0.2.6
    # via prompt-toolkit
websockets==11.0.3
    # via uvicorn
wrapt==1.15.0
    # via otter-grader
